pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "tk-db-utils"
version = "0.2.0"
description = "一个基于 SQLAlchemy 2.0 的高可复用数据库工具库，提供完整的 CRUD 操作、批量数据处理和高级数据库功能。"
readme = "README.md"
requires-python = ">=3.11"
license = {file = "LICENSE"}
authors = [
    {name = "tiankongzhise", email = "tiankongzhise@163.com"}
]
keywords = ["sqlalchemy", "database", "orm", "crud", "mysql", "postgresql", "sqlite"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Database",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
dependencies = [
    "pydantic>=2.11.4",
    "pymysql>=1.1.1",
    "python-dotenv>=1.1.1",
    "sqlalchemy>=2.0.40",
    "tk-base-utils>=0.2.5",
]

[project.urls]
Homepage = "https://github.com/tiankongzhise/tk-db-utils"
Repository = "https://github.com/tiankongzhise/tk-db-utils"
Documentation = "https://github.com/tiankongzhise/tk-db-utils#readme"
Issues = "https://github.com/tiankongzhise/tk-db-utils/issues"
